tried to make the animation for eating work but didn't find the solution yet.
Also fixed a bug when the chicken speak. since I wanted them to say their thing only once, I added a  variable to count the times, but then when the player approached other chicken they didn't talk again, so I fixed a condition for the distance and now it seems to work.