I added octaves on the grid
I separated the grid in 6 sections horizontally, and stored the id of the keys corresponding to those sections in different arrays.
then when the music play, i check to which array the current key played belongs to, and adjust the octave according to it.
also added some visuals to represent the octaves on the side of the grid